[C-01]: Move all components to component file
