+++ src/containers/Homepage/components/GroupDiv.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const GroupDiv = styled.div`
+  margin-top: 32px;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+`;
