+++ src/containers/Homepage/components/Layout.js
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const Layout = styled.div``;
