+++ src/containers/Homepage/components/SubjectDiv.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const SubjectDiv = styled.div`
+  margin-top: 32px;
+  display: block;
+  justify-content: center;
+  text-align: center;
+`;
