+++ src/containers/Homepage/components/SubjectImg.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const MathSubjectImg = styled.img`
+  alt: "Math Subject Image";
+  width: 60px;
+  height: 60px;
+  margin: 24px 32px 24px 32px;
+`;
+
+export const EngSubjectImg = styled.img`
+  alt: "English Subject Image";
+  width: 60px;
+  height: 60px;
+  margin: 24px 32px 24px 32px;
+`;
