+++ src/containers/Homepage/components/SubjectTitle.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+// Colors
+import { CHARCOAL } from "../../../global/const";
+
+export const SubjectTitle = styled.p`
+  font-family: Prompt;
+  font-weight: 600;
+  font-size: 24px;
+  line-height: 1.2em;
+  color: ${CHARCOAL};
+  margin: 40px 95px 40px 0px;
+`;
